+++ integration-tests/dataformats-json/src/main/java/org/apache/camel/quarkus/component/dataformats/json/JsonDataformatsRoute.java
@@ -19,7 +19,6 @@ package org.apache.camel.quarkus.component.dataformats.json;
 import java.lang.reflect.Type;
 import java.util.Arrays;
 import java.util.List;
-import java.util.TimeZone;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
@@ -30,20 +29,13 @@ import com.google.gson.reflect.TypeToken;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.gson.GsonDataFormat;
 import org.apache.camel.component.jackson.JacksonDataFormat;
-import org.apache.camel.component.jacksonxml.JacksonXMLDataFormat;
-import org.apache.camel.component.jacksonxml.ListJacksonXMLDataFormat;
 import org.apache.camel.component.johnzon.JohnzonDataFormat;
 import org.apache.camel.component.jsonb.JsonbDataFormat;
 import org.apache.camel.model.dataformat.JsonLibrary;
 import org.apache.camel.quarkus.component.dataformats.json.model.DummyObject;
 import org.apache.camel.quarkus.component.dataformats.json.model.ExcludeField;
-import org.apache.camel.quarkus.component.dataformats.json.model.MyModule;
 import org.apache.camel.quarkus.component.dataformats.json.model.PojoA;
 import org.apache.camel.quarkus.component.dataformats.json.model.PojoB;
-import org.apache.camel.quarkus.component.dataformats.json.model.TestJAXBPojo;
-import org.apache.camel.quarkus.component.dataformats.json.model.TestPojo;
-import org.apache.camel.quarkus.component.dataformats.json.model.TestPojoView;
-import org.apache.camel.quarkus.component.dataformats.json.model.Views;
 import org.apache.camel.spi.DataFormat;
 import org.apache.johnzon.mapper.reflection.JohnzonParameterizedType;
 
@@ -95,97 +87,6 @@ public class JsonDataformatsRoute extends RouteBuilder {
         from("direct:jacksonxml-unmarshal")
                 .unmarshal()
                 .jacksonxml(PojoA.class);
-
-        JacksonXMLDataFormat unmarshalTypeHeaderFormat = new JacksonXMLDataFormat();
-        unmarshalTypeHeaderFormat.setAllowUnmarshallType(true);
-        from("direct:jacksonxml-unmarshal-type-header").unmarshal(unmarshalTypeHeaderFormat);
-
-        JacksonDataFormat listFormat = new JacksonDataFormat(TestPojo.class);
-        listFormat.useList();
-        from("direct:jacksonxml-unmarshal-list").unmarshal(listFormat);
-
-        JacksonXMLDataFormat listSplitFormat = new JacksonXMLDataFormat(DummyObject.class);
-        listSplitFormat.useList();
-        from("direct:jacksonxml-unmarshal-listsplit").unmarshal(listSplitFormat).split(body()).to("mock:listsplit-result");
-
-        JacksonDataFormat format = new JacksonDataFormat();
-        from("direct:jacksonxml-marshal-includedefault").marshal(format).to("mock:marshal-includedefalut");
-
-        JacksonDataFormat marshalContentTypeformat = new JacksonDataFormat();
-        from("direct:jacksonxml-marshal-ct-yes").marshal(marshalContentTypeformat);
-        from("direct:jacksonxml-marshal-ct-yes2").marshal().json(JsonLibrary.Jackson);
-        JacksonDataFormat marshalContentTypeFormatNoHeader = new JacksonDataFormat();
-        marshalContentTypeFormatNoHeader.setContentTypeHeader(false);
-        from("direct:jacksonxml-marshal-ct-no").marshal(marshalContentTypeFormatNoHeader);
-
-        JacksonDataFormat jacksonMarshalFormat = new JacksonDataFormat();
-        from("direct:jacksonxml-marshal-in").marshal(jacksonMarshalFormat);
-        from("direct:jacksonxml-marshal-back").unmarshal(jacksonMarshalFormat).to("mock:reverse");
-        JacksonDataFormat prettyPrintDataFormat = new JacksonDataFormat();
-        prettyPrintDataFormat.setPrettyPrint(true);
-        from("direct:jacksonxml-marshal-inPretty").marshal(prettyPrintDataFormat);
-        from("direct:jacksonxml-marshal-backPretty").unmarshal(prettyPrintDataFormat).to("mock:reverse");
-        JacksonDataFormat formatPojo = new JacksonDataFormat(TestPojo.class);
-        from("direct:jacksonxml-marshal-inPojo").marshal(formatPojo);
-        from("direct:jacksonxml-marshal-backPojo").unmarshal(formatPojo).to("mock:reversePojo");
-
-        JacksonXMLDataFormat jacksonXmlMarshalAllowJmsTypeformat = new JacksonXMLDataFormat();
-        jacksonXmlMarshalAllowJmsTypeformat.setAllowJmsType(true);
-        from("direct:jacksonxml-marshal-allowjmstype-backPojo").unmarshal(jacksonXmlMarshalAllowJmsTypeformat)
-                .to("mock:allowjmstype-reversePojo");
-
-        JacksonXMLDataFormat jacksonXmlMarshalModuleformat = new JacksonXMLDataFormat();
-        jacksonXmlMarshalModuleformat.setInclude("NON_NULL");
-        jacksonXmlMarshalModuleformat.setModuleClassNames("org.apache.camel.quarkus.component.dataformats.json.model.MyModule");
-        from("direct:jacksonxml-marshal-module").marshal(jacksonXmlMarshalModuleformat).to("mock:jacksonxml-marshal-module");
-
-        from("direct:jacksonxml-marshal-concurrent-start").marshal().jacksonxml().to("log:marshalled")
-                .to("direct:jacksonxml-marshal-concurrent-marshalled");
-        from("direct:jacksonxml-marshal-concurrent-marshalled").unmarshal().jacksonxml(TestPojo.class)
-                .to("mock:jacksonxml-marshal-concurrent-result");
-
-        from("direct:jacksonxml-marshal-conversion").convertBodyTo(TestPojo.class);
-
-        from("direct:jacksonxml-unmarshal-listjackson").unmarshal(new ListJacksonXMLDataFormat(TestPojo.class))
-                .to("mock:jacksonxml-unmarshal-listjackson");
-
-        JacksonXMLDataFormat jacksonXmlJaxbAnnotationFormat = new JacksonXMLDataFormat();
-        from("direct:jacksonxml-jaxbannotation-in").marshal(jacksonXmlJaxbAnnotationFormat);
-        from("direct:jacksonxml-jaxbannotation-back").unmarshal(jacksonXmlJaxbAnnotationFormat)
-                .to("mock:jacksonxml-jaxbannotation-reverse");
-        JacksonXMLDataFormat jacksonXmlJaxbAnnotationformatPojo = new JacksonXMLDataFormat(TestJAXBPojo.class);
-        from("direct:jacksonxml-jaxbannotation-inPojo").marshal(jacksonXmlJaxbAnnotationformatPojo);
-        from("direct:jacksonxml-jaxbannotation-backPojo").unmarshal(jacksonXmlJaxbAnnotationformatPojo)
-                .to("mock:jacksonxml-jaxbannotation-reversePojo");
-
-        from("direct:jacksonxml-jsonview-inPojoAgeView").marshal().jacksonxml(TestPojoView.class, Views.Age.class);
-        from("direct:jacksonxml-jsonview-backPojoAgeView").unmarshal().jacksonxml(TestPojoView.class)
-                .to("mock:jacksonxml-jsonview-reversePojoAgeView");
-        from("direct:jacksonxml-jsonview-inPojoWeightView").marshal().jacksonxml(TestPojoView.class, Views.Weight.class);
-        from("direct:jacksonxml-jsonview-backPojoWeightView").unmarshal().jacksonxml(TestPojoView.class)
-                .to("mock:jacksonxml-jsonview-reversePojoWeightView");
-
-        this.getContext().getRegistry().bind("myJacksonModule", new MyModule());
-        JacksonXMLDataFormat jacksonXmlModuleRef = new JacksonXMLDataFormat();
-        jacksonXmlModuleRef.setInclude("NON_NULL");
-        jacksonXmlModuleRef.setModuleRefs("myJacksonModule");
-        from("direct:jacksonxml-moduleref-marshal").marshal(jacksonXmlModuleRef).to("mock:jacksonxml-moduleref-marshal");
-
-        JacksonXMLDataFormat jacksomXmlIncludeNotNullFormat = new JacksonXMLDataFormat();
-        jacksomXmlIncludeNotNullFormat.setInclude("NON_NULL");
-        from("direct:jacksonxml-include-non-null-marshal").marshal(jacksomXmlIncludeNotNullFormat)
-                .to("mock:jacksonxml-include-non-null-marshal");
-
-        JacksonXMLDataFormat jacksonXmlTypeHeaderNotAllowedFormat = new JacksonXMLDataFormat();
-        from("direct:jacksonxml-typeheader-not-allowed-backPojo").unmarshal(jacksonXmlTypeHeaderNotAllowedFormat)
-                .to("mock:jacksonxml-typeheader-not-allowed-reversePojo");
-
-        JacksonXMLDataFormat jacksonXmlDateTimezoneFormat = new JacksonXMLDataFormat();
-        TimeZone timeZone = TimeZone.getTimeZone("Africa/Ouagadougou");
-        jacksonXmlDateTimezoneFormat.setTimezone(timeZone);
-
-        from("direct:jacksonxml-datatimezone-in").marshal(jacksonXmlDateTimezoneFormat)
-                .to("mock:jacksonxml-datatimezone-result");
 
     }
